pyproject.toml
[project]
name = "musictl"
version = "0.1.0"
description = "Music control utility with wofi integration"
authors = [
    {name = "Sanek", email = "anton@shuvalov.info"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "pyrofi>=0.3.0,<0.4.0",
    "pyyaml>=6.0.2,<7.0.0",
    "dbus-python>=1.4.0,<2.0.0"
]

[project.scripts]
musictl = "musictl.__main__:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
